Switch some int types

+++ src/prototypes/common.h
@@ -26,7 +26,7 @@ void get_kernel(char *, uint8_t num);
 void get_packs(char *);
 void get_taim(char *);
 void get_uptime(char *);
-void check_fan_vals(char *, uint_fast16_t *, uint8_t iterz);
+void check_fan_vals(char *, uint_fast16_t *, uint_fast16_t iterz);
 void split_n_index(char *);
 
 #if !defined(HAVE_SENSORS_SENSORS_H) && !defined(__OpenBSD__)
